Flytt funksjonalitet for å oppdatere kompetanser til egen klasse
